requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
groq==0.11.0
python-multipart==0.0.19
python-dotenv
motor>=3.3.0
pymongo>=4.6.0
requests
sentence-transformers==3.0.1
safetensors==0.4.3
tokenizers==0.19.1
transformers==4.43.3